r/aws_redshiftserverless_workgroup: Workgroup may be 'AVAILABLE' when deletion starts.
